Move InspectDatabase to database_ext.go
